feat: new user signup persist
